New types, fix imports
