+++ app/ts/components/Transactions.tsx
@@ -3,7 +3,7 @@ import { utils } from 'ethers'
 import { JSXInternal } from 'preact/src/jsx.js'
 import { createBundleTransactions, } from '../library/bundleUtils.js'
 import { FlashbotsBundleTransaction } from '../library/flashbots-ethers-provider.js'
-import { AppSettings, BlockInfo, BundleState, Signers } from '../library/types.js'
+import { AppSettings, BlockInfo, BundleState, Signers } from '../types/types.js'
 import { MEV_RELAY_GOERLI } from '../constants.js'
 import { ProviderStore } from '../library/provider.js'
 import { Button } from './Button.js'
